Move output format for file events
