+++ tests/cmds/test_securitydata.py
@@ -12,12 +12,8 @@ from tests.conftest import get_test_date_str
 from code42cli import errors
 from code42cli import PRODUCT_NAME
 from code42cli.cmds.search.cursor_store import FileEventCursorStore
-from code42cli.cmds.securitydata import file_events_output_format
-from code42cli.cmds.securitydata_output_formats import to_cef
 from code42cli.main import cli
-from code42cli.output_formats import to_dynamic_csv
-from code42cli.output_formats import to_formatted_json
-from code42cli.output_formats import to_json
+
 
 BEGIN_TIMESTAMP = 1577858400.0
 END_TIMESTAMP = 1580450400.0
@@ -725,23 +721,3 @@ def test_saved_search_list_with_format_option_does_not_return_when_response_is_e
         cli, ["security-data", "saved-search", "list", "-f", "csv"], obj=cli_state
     )
     assert "Name,Id" not in result.output
-
-
-def test_file_events_output_format_returns_to_dynamic_csv_function_when_csv_option_is_passed():
-    extraction_output_format_function = file_events_output_format(None, None, "CSV")
-    assert id(extraction_output_format_function) == id(to_dynamic_csv)
-
-
-def test_file_events_output_format_returns_to_formatted_json_function_when_json__option_is_passed():
-    format_function = file_events_output_format(None, None, "JSON")
-    assert id(format_function) == id(to_formatted_json)
-
-
-def test_file_events_output_format_returns_to_json_function_when_raw_json_format_option_is_passed():
-    format_function = file_events_output_format(None, None, "RAW-JSON")
-    assert id(format_function) == id(to_json)
-
-
-def test_file_events_output_format_returns_to_cef_function_when_cef_format_option_is_passed():
-    format_function = file_events_output_format(None, None, "CEF")
-    assert id(format_function) == id(to_cef)
